+++ packages/media/media-image/examples/0-basic.tsx
@@ -1,57 +1,27 @@
 import * as React from 'react';
 import { Component } from 'react';
 import FieldText from '@atlaskit/field-text';
-import { Auth, isClientBasedAuth } from '@atlaskit/media-core';
 import {
   genericFileId,
-  defaultParams,
   defaultCollectionName,
-  StoryBookAuthProvider,
   createStorybookContext,
 } from '@atlaskit/media-test-helpers';
+import Spinner from '@atlaskit/spinner';
+
 import { MediaImage } from '../src';
 
 export interface ExampleProps {}
 
 export interface ExampleState {
-  token: string;
   imageId: string;
-  collectionName: string;
-  clientId: string;
-  baseUrl: string;
 }
 
 const context = createStorybookContext();
 
 class Example extends Component<ExampleProps, ExampleState> {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      token: '',
-      imageId: genericFileId.id,
-      collectionName: defaultCollectionName,
-      clientId: defaultParams.clientId,
-      baseUrl: defaultParams.baseUrl,
-    };
-  }
-
-  componentDidMount() {
-    const authProvider = StoryBookAuthProvider.create(false);
-    authProvider({ collectionName: defaultCollectionName }).then(
-      (auth: Auth) => {
-        this.setState({
-          token: auth.token,
-        });
-
-        if (isClientBasedAuth(auth)) {
-          this.setState({
-            clientId: auth.clientId,
-          });
-        }
-      },
-    );
-  }
+  state: ExampleState = {
+    imageId: genericFileId.id,
+  };
 
   onIdChange = e => {
     this.setState({
@@ -59,37 +29,8 @@ class Example extends Component<ExampleProps, ExampleState> {
     });
   };
 
-  onCollectionChange = e => {
-    this.setState({
-      collectionName: e.target.value,
-    });
-  };
-
-  onTokenChange = e => {
-    this.setState({
-      token: e.target.value,
-    });
-  };
-
-  onClientIdChange = e => {
-    this.setState({
-      clientId: e.target.value,
-    });
-  };
-
-  onBaseUrlChange = e => {
-    this.setState({
-      baseUrl: e.target.value,
-    });
-  };
-
   render() {
-    const { token, imageId, collectionName, clientId, baseUrl } = this.state;
-    const apiConfig = {
-      token,
-      clientId,
-      baseUrl,
-    };
+    const { imageId } = this.state;
 
     return (
       <div>
@@ -110,30 +51,6 @@ class Example extends Component<ExampleProps, ExampleState> {
             value={imageId}
             onChange={this.onIdChange}
           />
-          <FieldText
-            label="Collection name"
-            placeholder="Collection name..."
-            value={collectionName}
-            onChange={this.onCollectionChange}
-          />
-          <FieldText
-            label="Token"
-            placeholder="Token..."
-            value={token}
-            onChange={this.onTokenChange}
-          />
-          <FieldText
-            label="Client id"
-            placeholder="Client id..."
-            value={clientId}
-            onChange={this.onClientIdChange}
-          />
-          <FieldText
-            label="Service host"
-            placeholder="Service host..."
-            value={baseUrl}
-            onChange={this.onBaseUrlChange}
-          />
         </div>
         <div
           style={{
@@ -145,8 +62,9 @@ class Example extends Component<ExampleProps, ExampleState> {
           <MediaImage
             id={imageId}
             context={context}
-            collectionName={collectionName}
+            collectionName={defaultCollectionName}
             width={300}
+            placeholder={<Spinner />}
           />
         </div>
       </div>