+++ packages/core/navigation-next/src/components/ContentNavigation/types.js
@@ -4,11 +4,8 @@ import type { ComponentType } from 'react';
 
 export type ContentNavigationProps = {
   container?: ?ComponentType<{}>,
-  isDragging: boolean,
   isPeekHinting: boolean,
   isPeeking: boolean,
-  onOverlayClick?: Event => void,
+  isVisible: boolean,
   product: ComponentType<{}>,
-  transitionState: 'entered' | 'entering' | 'exited' | 'exiting',
-  width: number,
 };