+++ packages/media/media-editor/docs/0-intro.tsx
@@ -1,5 +1,4 @@
-import * as React from 'react';
-import { md, Example, Props } from '@atlaskit/docs';
+import { md } from '@atlaskit/docs';
 
 export default md`
   # MediaFilmstrip