+++ packages/search/global-search/src/__tests__/unit/api/JiraClient.test.ts
@@ -8,7 +8,7 @@ import {
 
 const url = 'https://www.example.jira.dev.com/';
 const cloudId = 'cloudId';
-const PATH = '/rest/internal/2/productsearch/recent';
+const PATH = 'rest/internal/2/productsearch/recent';
 describe('JiraClient', () => {
   let requestSpy;
   let jiraClient;