+++ packages/media/media-card/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @atlaskit/media-image
+# @atlaskit/media-card
 
 ## 34.0.4
 - [patch] Refetch image when the dimensions change [a0475c2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a0475c2)